requirements.txt
google-cloud-storage
google-cloud-bigquery
python-dotenv~=1.0.1
flask~=3.0.3
